requirements.txt
paramiko==2.4.2
PGPy==0.5.2
psycopg2-binary==2.8.3
pytest==5.0.1
pytest-cov==2.7.1
-e git+https://github.com/MoveOnOrg/python-actionkit.git@public-master#egg=python-actionkit
python-coveralls==2.9.3
git+https://github.com/sreynen/pywell@main#egg=pywell
slackweb==1.0.5
zappa==0.48.2
